[*] Move DateUtils functions back to EmployeeInfoEntity
